pyproject.toml
[project]
name = "openapi-rest-proxy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.6",
    "httpx>=0.28.1",
    "pytest>=8.3.4",
    "pyyaml>=6.0.2",
    "uvicorn>=0.34.0",
]
